Create README.md
Im working im a team of 2 to build an online addiction recovery business called the addiction coach

Our research indicates that there is a growing demand for addiction related services, further accelerated by the COVID pandemic, and that the existing support network is woefully ineffective. For example we suspect that the rehab failure rate is alarmingly high. The addiction coach solves this problem by offering an online coaching service that actually works in 2022

In the last 6 months weve worked hard to validate our MVP and we are certain there is a market demand of at least 100000 people

We have strong domain expertise: I am the ex addict myself with effective and unusual modern day solutions to the problem, and my co founder has 20 years experience in tech based testing management

Thanks for taking the time and id love to tell you more!!!!